ReadUserClass new method to execute user input
need to fix get and delete method in history class

+++ DataStructureTask/Observation.cs
@@ -6,10 +6,10 @@ namespace DataStructureTask
     public class Observation
     {
 
-        public int TimeStamp {get; set;}
+        public long TimeStamp {get; set;}
         public string  Data { get; set; }
 
-        public Observation(int timeStamp, string data)
+        public Observation(long timeStamp, string data)
         {
             this.TimeStamp = timeStamp;
             this.Data = data;
